pyproject.toml
[tool.poetry]
name = "dialog-lib"
version = "0.0.2.3"
description = ""
authors = ["Talkd.AI <foss@talkd.ai>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "<0.3.0"
SQLAlchemy = "^2.0.29"
click = "^8.1.7"
pgvector = "^0.2.5"
langchain-openai = "^0.1.8"
psycopg2-binary = "^2.9.9"
langchain-postgres = "0.0.7"
langchain-community = "^0.2.5"
langchain-anthropic = "^0.1.11"
bs4 = "^0.0.2"
responses = "^0.25.0"
aioresponses = "^0.7.6"
pytest-asyncio = "^0.23.6"
gspread = "^6.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
black = "^24.3.0"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
dialog = "dialog_lib.manage:main"

[tool.poetry.scripts]
dialog = "dialog_lib.manage:main"